05_modules_packages/example_project/requirements.txt
certifi==2026.2.25
charset-normalizer==2.0.12
click==8.1.7
Flask==2.3.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
loguru==0.6.0
MarkupSafe==3.0.3
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
six==1.17.0
typing_extensions==4.15.0
urllib3==1.26.20
Werkzeug==3.1.7
